New Bomb Page refactoring
